Some example code for using the autocomplete LC
autocompleteSobject is an LC that implements something much like the
native Salesforce "lookup" field.

autocompleteHHAddress is an LC that customizes the autocomplete LC to
show a list of contacts, with their addresses, an icon, and a button to
take action.  It re-uses the dataProviderSobject LC from the other
example, but it would probably be better to write a custom provider for
a real implementation.